Added more functions to utils
