+++ facs/base/facs.py
@@ -17,7 +17,14 @@ from .needs import Needs
 from .location_types import building_types_dict, building_types
 from .house import House
 from .location import Location
-from .utils import probability, get_random_int, out_files, calc_dist
+from .utils import (
+    probability,
+    get_random_int,
+    out_files,
+    calc_dist,
+    write_log_headers,
+    check_vac_eligibility,
+)
 from .mpi import MPIManager
 
 log_prefix = "."
@@ -28,29 +35,6 @@ num_infections_today = 0
 num_hospitalisations_today = 0
 num_deaths_today = 0
 num_recoveries_today = 0
-
-
-def write_log_headers(rank):
-    out_inf = out_files.open("{}/covid_out_infections_{}.csv".format(log_prefix, rank))
-    print("#time,x,y,location_type,rank,incubation_time", file=out_inf, flush=True)
-    out_inf = out_files.open(
-        "{}/covid_out_hospitalisations_{}.csv".format(log_prefix, rank)
-    )
-    print("#time,x, y,age", file=out_inf, flush=True)
-    out_inf = out_files.open("{}/covid_out_deaths_{}.csv".format(log_prefix, rank))
-    print("#time,x,y,age", file=out_inf, flush=True)
-    out_inf = out_files.open("{}/covid_out_recoveries_{}.csv".format(log_prefix, rank))
-    print("#time,x,y,age", file=out_inf, flush=True)
-
-
-def check_vac_eligibility(a):
-    if (
-        a.status == "susceptible"
-        and a.symptoms_suppressed == False
-        and a.antivax == False
-    ):
-        return True
-    return False
 
 
 class Ecosystem:
